Move content of the config sub-pkg into config.go
