+++ pkg/controller/apmserver/config.go
@@ -2,17 +2,23 @@
 // or more contributor license agreements. Licensed under the Elastic License;
 // you may not use this file except in compliance with the Elastic License.
 
-package config
+package apmserver
 
 import (
 	"fmt"
 	"path"
 	"path/filepath"
 
+	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+
 	apmv1 "github.com/elastic/cloud-on-k8s/pkg/apis/apm/v1"
 	commonv1 "github.com/elastic/cloud-on-k8s/pkg/apis/common/v1"
+	"github.com/elastic/cloud-on-k8s/pkg/controller/apmserver/labels"
+	"github.com/elastic/cloud-on-k8s/pkg/controller/apmserver/name"
 	"github.com/elastic/cloud-on-k8s/pkg/controller/association"
 	"github.com/elastic/cloud-on-k8s/pkg/controller/common/certificates"
+	"github.com/elastic/cloud-on-k8s/pkg/controller/common/reconciler"
 	"github.com/elastic/cloud-on-k8s/pkg/controller/common/settings"
 	"github.com/elastic/cloud-on-k8s/pkg/utils/k8s"
 )
@@ -25,12 +31,42 @@ const (
 	CertificatesDir = "config/elasticsearch-certs"
 
 	APMServerHost        = "apm-server.host"
-	APMServerSecretToken = "apm-server.secret_token"
+	APMServerSecretToken = "apm-server.secret_token" // nolint
 
 	APMServerSSLEnabled     = "apm-server.ssl.enabled"
 	APMServerSSLKey         = "apm-server.ssl.key"
 	APMServerSSLCertificate = "apm-server.ssl.certificate"
+
+	ApmCfgSecretKey = "apm-server.yml" // nolint
 )
+
+// ReconcileConfig reconciles the configuration of the APM server: it first creates the configuration from the APM
+// specification and then reconcile the underlying secret.
+func ReconcileConfig(client k8s.Client, as *apmv1.ApmServer) (corev1.Secret, error) {
+	// Create a new configuration from the APM object spec.
+	cfg, err := NewConfigFromSpec(client, as)
+	if err != nil {
+		return corev1.Secret{}, err
+	}
+
+	cfgBytes, err := cfg.Render()
+	if err != nil {
+		return corev1.Secret{}, err
+	}
+
+	// ReconcileConfig the configuration in a secret
+	expectedConfigSecret := corev1.Secret{
+		ObjectMeta: metav1.ObjectMeta{
+			Namespace: as.Namespace,
+			Name:      name.Config(as.Name),
+			Labels:    labels.NewLabels(as.Name),
+		},
+		Data: map[string][]byte{
+			ApmCfgSecretKey: cfgBytes,
+		},
+	}
+	return reconciler.ReconcileSecret(client, expectedConfigSecret, as)
+}
 
 func NewConfigFromSpec(c k8s.Client, as *apmv1.ApmServer) (*settings.CanonicalConfig, error) {
 	specConfig := as.Spec.Config
